Organize taxonomy code
