+++ site/gatsby-site/src/templates/cite.js
@@ -409,46 +409,7 @@ export const pageQuery = graphql`
     }
 
     allMongodbAiidprodClassifications(filter: { incident_id: { eq: $incident_id } }) {
-      nodes {
-        incident_id
-        id
-        namespace
-        classifications {
-          Annotation_Status
-          Annotator
-          Ending_Date
-          Beginning_Date
-          Full_Description
-          Harm_Distribution_Basis
-          Intent
-          Location
-          Named_Entities
-          Near_Miss
-          Quality_Control
-          Reviewer
-          Severity
-          Short_Description
-          Technology_Purveyor
-          AI_Applications
-          AI_System_Description
-          AI_Techniques
-          Data_Inputs
-          Finacial_Cost
-          Harm_Distribution_Basis
-          Harm_Type
-          Infrastructure_Sectors
-          Laws_Implicated
-          Level_of_Autonomy
-          Lives_Lost
-          Nature_of_End_User
-          Physical_System
-          Problem_Nature
-          Public_Sector_Deployment
-          Relevant_AI_functions
-          Sector_of_Deployment
-          System_Developer
-        }
-      }
+      ...ClassificationFields
     }
 
     allMongodbAiidprodTaxa(filter: { namespace: { in: $taxonomy_namespace_array } }) {
